DbManager.insertRevision() should be taking in a long for the delta, not an int
git-svn-id: http://fbms.googlecode.com/svn/trunk@170 c375ce47-9a4c-63b9-7dad-5ffa0e8ccef7

+++ main/src/cmpt370/fbms/DbManager.java
@@ -183,7 +183,7 @@ public class DbManager
 		return null;
 	}
 
-	public static void insertRevision(Path file, String diff, long timestamp, int delta)
+	public static void insertRevision(Path file, String diff, long timestamp, long delta)
 	{
 
 	}
